+++ bin/report_single_sample.py
@@ -44,10 +44,9 @@ def make_result_qr_code(
         border=0)
 
     res = '|'.join([antibiotic for antibiotic in resistance['resistant']])
-    sus = '|'.join([antibiotic for antibiotic in resistance['susceptible']])
 
     level = resistance['resistance_level']
-    qr_string = f"""{sample_id},{barcode},{level},RES;{res},SUS;{sus}"""
+    qr_string = f"""{sample_id},{barcode},{level},RES;{res}"""
 
     qr.add_data(qr_string)
 
@@ -110,9 +109,9 @@ def make_assay_section(
     section._add_item("""</tr><thead><tr>""")
 
     values = dict(
-        platform="GridION",
+        platform="",
         version=f"{revision} ({commit})",
-        reference="AL123456.3",
+        reference="NC_000962.3",
         barcode=barcode
     )
 
@@ -214,17 +213,19 @@ def make_drug_section(
     section = report_doc.add_section()
 
     section._add_item(
-        f"""{make_header(data['drug_susceptibility']["title"])}""")
+        f"""{make_header(data['sections']['drug_susceptibility']["title"])}""")
 
     make_final_result_section(
-        section, data['final'], antibiotics, resistance, sample_id)
+        section, data['sections']['final'], antibiotics, resistance, sample_id)
 
     section._add_item(f"""
         <div class="row">
-            <div class="col">{data['drug_susceptibility']['blurb']}</div>
+            <div class="col">
+                {data['sections']['drug_susceptibility']['blurb']}
+            </div>
         <div class="col">""")
 
-    for k, v in data['drug_susceptibility']['result'].items():
+    for k, v in data['sections']['drug_susceptibility']['result'].items():
         checked = ''
         if k == resistance["resistance_level"]:
             checked = " checked"
@@ -261,13 +262,13 @@ def make_drug_section(
             <tr>
                 <td colspan=4 class="">
                     <strong>
-                        {data['drug_susceptibility']["line"][line]}
+                        {data['sections']['drug_susceptibility']["line"][line]}
                     </strong>
                 </td>
         </tr>""")
 
         for status in ['resistant', 'susceptible']:
-            interpretation = status.capitalize()
+            interpretation = data['misc_language'][status]
             for antibiotic in resistance[status]:
                 if line != resistance[status][antibiotic]['line']:
                     continue
@@ -430,7 +431,7 @@ def main():
     if controls is True:
 
         make_drug_section(
-            report_doc, canned_text['sections'], canned_text['antibiotics'],
+            report_doc, canned_text, canned_text['antibiotics'],
             resistance, args.sample_id, args.barcode)
 
     make_authorisation_section(